Fix:コメント箇所を修正
修正前プルリクエスト:https://github.com/okateru0217/pomodoro-app/pull/16
