--- .gitignore
@@ -1,3 +1,0 @@
-UserInterfaceState.xcuserstate
-.DS_Store
-*.swp
